startign to add player state in server and send to client integrating stuff from tinywarriors for ui

+++ Server/Assets/Scripts/Server/Components.cs
@@ -11,7 +11,7 @@ namespace Server
     
     public struct PlayerController : IComponentData
     {
-        public int gold;
+        public ushort gold;
     }
 
     public struct IdleAction : IComponentData
